Fix issues caused by refactoring
